Rename awslib -> s3lib.
