+++ lib/awslib/s3.go
@@ -1,4 +1,4 @@
-package s3lib
+package awslib
 
 import (
 	"context"
